inter-domain interface
